Modified the skills and user controllers and routers

+++ controllers/skill_controller.js
@@ -13,7 +13,7 @@ export const addSkill = async (req, res) => {
         // Find the user with the ID that was passed when the skill was being created
         console.log('userId', req.session.user.id)
 
-        const userSessionId = req.session.user.id
+        const userSessionId = req.session.user.id;
 
         const user = await userModel.findById(userSessionId);
         if(!user) {
@@ -25,7 +25,7 @@ export const addSkill = async (req, res) => {
         
 
         // if you find the user, push the skill id you just created inside
-        user.skill.push(skill._id);
+        user.skills.push(skill._id);
 
         // and save the user now with skillId
         await user.save();
@@ -39,17 +39,26 @@ export const addSkill = async (req, res) => {
 }
 
 // Function to update a skill
-export const patchSkill = async (req, res, next) => {
+export const patchSkill = async (req, res) => {
     try {
         const {error, value} = skillsSchema.validate(req.body)
         if(error) {
             return res.status(400).json(error.details[0].message)
         }
-        const updatedSkill = await skillModel.findByIdAndUpdate(req.params.id, req.body, {new: true})
+
+        const userSessionId = req.sesion.user.id;
+        const user = await userModel.findById(userSessionId);
+        if(!user) {
+            return res.status(404).send("User not found")
+        }
+        const updatedSkill = await skillModel.findByIdAndUpdate(req.params.id, value, {new: true});
+        if (!skill) {
+            return res.status(404).send('Skill not found');
+        }
         // Return response
-        res.status(200).json(updatedSkill)
+        res.status(200).json({Skill: updatedSkill})
     } catch (error) {
-        next(error)
+        return res.status(500).json({error})
     }
 }
 
@@ -63,7 +72,7 @@ export const getSkills = async (req, res) => {
             return res.status(404).send('No skills added')
         }
         // Return all skills
-        res.status(200).json({skills: allSkills})
+        res.status(200).json({Skills: allSkills})
     } catch (error) {
         return res.status(500).send('Internal error')
     }
@@ -82,9 +91,22 @@ export const getOneSkill = async (req, res, next) => {
 // Function to delete a skill with a unique ID
 export const deleteSkill = async (req, res, next) => {
     try {
-        const deleteData = await skillModel.findByIdAndDelete(req.params.id)
+
+        const userSessionId = req.session.user.id;
+        const user = await userModel.findById(userSessionId);
+
+        if (!user) {
+            return res.staus(404).send("User not found");
+        }
+        const skill = await skillModel.findByIdAndDelete(req.params.id)
+        if(!skill) {
+            return res.status(404).send("Skill not found");
+        }
+
+        user.skills.pull(req.params.id);
+        await user.save();
         res.status(200).send(`Skill with ID ${req.params.id} has been deleted`)
     } catch (error) {
-        next(error)
+        return res.status(500).json({error})
     }
-}
+};
